refactor(k3): move K3PagedKv into its own executor module
Pure code motion: the paged latent cache (struct, page allocator, block
tables, append/adopt paths) and its two geometry constants move from
executor/buffers.rs to executor/paged_kv.rs; copy_rows stays in buffers
as pub(super). No behavior change.
