+++ pegainfer-k3/src/executor/step.rs
@@ -65,17 +65,17 @@ use pegainfer_kernels::tensor::DeviceMatrix;
 
 use super::buffers::K3_KDA_FUSED;
 use super::buffers::K3_KDA_WSM_PADDED;
-use super::buffers::K3_KV_PAGE_TOKENS;
 use super::buffers::K3_MLA_FUSED;
-use super::buffers::K3_MLA_LATENT_ROW;
 use super::buffers::K3LayerState;
-use super::buffers::K3PagedKv;
 use super::buffers::K3Scratch;
 use super::buffers::K3StatePool;
 use super::buffers::parity_pair;
 use super::gemm::K3PartialSpan;
 use super::gemm::k3_gemm_full;
 use super::gemm::k3_gemm_partial;
+use super::paged_kv::K3_KV_PAGE_TOKENS;
+use super::paged_kv::K3_MLA_LATENT_ROW;
+use super::paged_kv::K3PagedKv;
 use crate::config::K3_ATTN_INNER;
 use crate::config::K3_DENSE_INTERMEDIATE;
 use crate::config::K3_EXPERT_INTERMEDIATE;
